refactor: refactor top app bar usage in the screens
