+++ design_system/src/main/java/com/baghdad/design_system/component/appBar/BasicTopAppBar.kt
@@ -1,5 +1,6 @@
 package com.baghdad.design_system.component.appBar
 
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -19,6 +20,7 @@ internal fun BasicTopAppBar(
             .fillMaxWidth()
             .padding(horizontal = 16.dp),
         verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.SpaceBetween
     ) {
         content()
     }
