+++ design_system/src/main/java/com/baghdad/design_system/component/appBar/TopAppBar.kt
@@ -18,7 +18,7 @@ fun TopAppBar(
     modifier: Modifier = Modifier,
     onGoBackClick: (() -> Unit)? = null,
     screenTitle: String? = null,
-    content: @Composable RowScope.() -> Unit,
+    content: (@Composable RowScope.() -> Unit)? = null
 ) {
 
     BasicTopAppBar(
@@ -43,7 +43,7 @@ fun TopAppBar(
                 modifier = Modifier.weight(1f)
             )
         }
-        content()
+        content?.let { it() }
     }
 }
