+++ ui/src/main/java/com/baghdad/ui/feature/trendingTvShow/TrendingTvShowScreen.kt
@@ -24,7 +24,6 @@ import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
 import com.baghdad.design_system.component.Scaffold
 import com.baghdad.design_system.component.SnackBar
-import com.baghdad.design_system.component.Text
 import com.baghdad.design_system.component.WavyLoadingIndicator
 import com.baghdad.design_system.component.appBar.TopAppBar
 import com.baghdad.design_system.theme.Theme
@@ -111,15 +110,7 @@ fun TrendingTvShowContent(
                 onGoBackClick = {
                     listener.onBackIconClick()
                 },
-                content = {
-                    Text(
-                        text = stringResource(R.string.trending_tv_shows),
-                        style = Theme.typography.title.large,
-                        color = Theme.color.title,
-                        modifier = Modifier
-                            .padding(start = 8.dp)
-                    )
-                }
+                screenTitle = stringResource(R.string.trending_tv_shows),
             )
         }
     ) {
