+++ ui/src/main/res/values-ar/strings.xml
@@ -57,7 +57,7 @@
     <string name="save_to_list">حفظ الى القائمه</string>
     <string name="guests_of_honor">ضيوف الشرف</string>
     <string name="season_template">موسم %1$s</string>
-    <string name="trending_tv_shows">Trending TV shows</string>
+    <string name="trending_tv_shows">المسلسلات الرائجة</string>
     <string name="movies">أفلام</string>
     <string name="tv_shows">مسلسلات</string>
     <string name="what_do_you_want_to_watch">ماذا تريد أن تشاهد؟</string>
@@ -68,4 +68,6 @@
     <string name="continue_watching">متابعة المشاهدة</string>
     <string name="upcoming">قادم</string>
     <string name="top_rating_item">Top rating item</string>
+    <string name="trending_people">الممثلون الرائجون</string>
+
 </resources>
